Extract into client component
